+++ R/md_stock1_.R
@@ -1,8 +1,8 @@
 #' dataset of shanghai composite index
 #'
-#' The daily historical Shanghai Composite Index from the beginning of the Index to 17th Sept 2020
+#' The daily historical Shanghai Composite Index from the beginning of the Index to Sept 1, 2020
 #'
-#' @format A data frame with 7518 rows and 15 variables:
+#' @format A data frame with 7506 rows and 15 variables:
 #' \describe{
 #'   \item{symbol}{stock ticker symbol}
 #'   \item{name}{stock ticker name}
@@ -35,7 +35,7 @@
 #' @param from the start date. Default is NULL.
 #' @param to the end date. Default is current system date.
 #' @param adjust whether to adjust the OHLC prices. If it is NULL or FALSE, return the original data. Default is FALSE. 
-#' For the yahoo data, the adjustment is based on 'close_adj' column; for the 163 data, the adjustment is based on 'close' and 'close_prev' columns.
+#' For the yahoo data, the adjustment is based on the close_adj; for the 163 data, the adjustment is based on the cumulative products of close/close_prev.
 #' @param print_step A non-negative integer. Print symbol name by each print_step iteration. Default is 1L.
 #' @param ... Additional parameters.
 #' 
@@ -85,6 +85,13 @@
 #' dt_spot4 = md_stock(symbol = c('a', 'b', 'index', 'fund'), source = '163', 
 #'   type = 'spot', show_tags = TRUE)
 #' 
+#' # Example IV
+#' # query company information, including profile, IPO, structure of income, structure of employee
+#' dt_info = md_stock('600036', type = 'info')
+#' 
+#' # query structure of income in history only
+#' dt_info2 = md_stock('600036', type = 'info', str_income_hist = TRUE)
+#' 
 #' }
 #' 
 #' @export
@@ -96,10 +103,10 @@ md_stock = function(symbol, source = "yahoo", type='history', freq = "daily", da
         check_internet('www.yahoo.com')
     }
     # arguments
-    type = check_arg(type, c('history', 'spot', 'adjfactor'))
+    type = check_arg(type, c('history', 'spot', 'adjfactor', 'info'))
     
     source = check_arg(as.character(source), c('yahoo','163'), default = 'yahoo')
-    if (type == 'spot') source = '163'
+    if (type %in% c('spot', 'info')) source = '163'
     
     syb = tolower(symbol)
     ## remove NAs from the yahoo data