+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * fixed mutual bugs in the functions such as get_fromto, md_stock_divsplit1_163 etc.
 * modified stock adjust method for 163 data
+* md_stock function supports to query listed company information, such as profile, IPO, structure of income, structure of employee
 
 # pedquant 0.1.6